# Conflicts:
#	src/main/java/com/github/vaerys/main/Constants.java

+++ src/main/java/com/github/vaerys/tags/cctags/TagDelCall.java
@@ -1,9 +1,11 @@
 package com.github.vaerys.tags.cctags;
 
+import com.github.vaerys.handlers.GuildHandler;
 import com.github.vaerys.masterobjects.CommandObject;
 import com.github.vaerys.enums.TagType;
 import com.github.vaerys.handlers.RequestHandler;
 import com.github.vaerys.templates.TagObject;
+import sx.blah.discord.handle.obj.Permissions;
 
 public class TagDelCall extends TagObject {
 
@@ -13,7 +15,9 @@ public class TagDelCall extends TagObject {
 
     @Override
     public String execute(String from, CommandObject command, String args) {
-        RequestHandler.deleteMessage(command.message.get());
+        if (GuildHandler.testForPerms(command,Permissions.MANAGE_MESSAGES)) {
+            RequestHandler.deleteMessage(command.message.get());
+        }
         return removeAllTag(from);
     }
